+++ src/App.js
@@ -1,31 +1,14 @@
 import './App.css'
-import { useState } from 'react'
+import Input from './components/Input'
+import ToDoList from './components/ToDoList'
+
 
 const list = document.getElementsByClassName('list')
 
-function App(props) {
-  const [text, setText] = useState('')
-
-  
- 
-  function createNewTask(event) {
-    event.preventDefault()
-  }
-
-  return <div className="App">
-    <div className='inputBlock'>
-      <form onSubmit={createNewTask}>
-        <h1>Todo app React</h1>
-        <input className='inputField'  onChange={(e) => setText(e.target.value)} ></input>
-        <button className='taskButton' type='submit'>Add task</button>  
-      </form>   
-    </div>
-
-    <ul className='list'>
-      <li>Bless the god imperror</li>
-      <li>Become an Inqusitor</li>
-      <li>{text}</li>
-    </ul>
+function App() {
+  return <div className='App'>
+    <Input />
+    <ToDoList />
   </div>
